refactor P2PBlobTransferManagerFactory to BlobTransferManagerBuilder
